backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0.post1
python-multipart==0.0.6
pandas==2.2.0
motor==3.3.2
pymongo==4.6.1
boto3==1.34.34
python-jose==3.3.0
httpx==0.26.0
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1
apscheduler